fix
Fix the direction from northeast to east, because the cave entrance is on the east an also Cithurn originay sed:
andors-trail-master\AndorsTrail\res\raw\conversationlist_graveyard1.json
"Thank you. You will have to pass through the forest to reach an opening to the cave where you can enter. The opening is roughly east of my home."